feat(store,ui): move submit layer into submit/ sub-packages + fix tests
Moves SubmitState, SubmitHandler, SubmitStateExtensions into
template.core.base.store.submit and SubmitProgressOverlay,
SubmitResultHandler into template.core.base.ui.submit — completing the
re-org started in the previous commit.

Test fixes:
- SubmitHandlerTest: use `this` (TestScope) instead of backgroundScope
  so StandardTestDispatcher's scheduler advances the handler's coroutines
  via advanceUntilIdle()
- T2/T10: replace RuntimeException("IOException: ...") with a local
  class IOException so categorize() detects the network class name,
  matching the pattern established in ErrorCategoryTest
