+++ core-base/store/src/commonMain/kotlin/template/core/base/store/submit/SubmitStateExtensions.kt
@@ -7,7 +7,9 @@
  *
  * See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
  */
-package template.core.base.store
+package template.core.base.store.submit
+
+import template.core.base.store.error.ErrorCategory
 
 /**
  * Extension properties for [SubmitState]. Mirror the style of [ScreenStateExtensions].
